+++ packages/firebase_ai/firebase_ai/lib/src/firebaseai_version.dart
@@ -13,4 +13,4 @@
 // limitations under the License.
 
 /// generated version number for the package, do not manually edit
-const packageVersion = '0.3.0';
+const packageVersion = '0.3.0+2';