+++ packages/firebase_auth/firebase_auth_web/lib/src/firebase_auth_version.dart
@@ -13,4 +13,4 @@
 // limitations under the License.
 
 /// generated version number for the package, do not manually edit
-const packageVersion = '6.4.0';
+const packageVersion = '6.5.1';